+++ layout/columns1.php
@@ -46,7 +46,7 @@ echo $OUTPUT->doctype() ?>
         <div class="container-fluid">
             <?php echo $OUTPUT->navbar_button(); ?>
             <?php echo $OUTPUT->user_menu(); ?>
-            <?php echo $OUTPUT->navbar_plugin_output(); ?>
+            <?php if ($CFG->branch >= 32) {echo $OUTPUT->navbar_plugin_output();} ?>
             <?php echo $OUTPUT->search_box(); ?>
             <div class="nav-collapse collapse">
                 <?php echo $OUTPUT->custom_menu(); ?>