+++ CHANGELOG.md
@@ -10,4 +10,4 @@ All notable changes to this project will be documented in this file.
 ## [0.1.0] - 2017-11-05
 ### Added
 - Initial public ALPHA release on GitHub.
-- Plugin officially compatible and tested with Moodle 2.5 to 3.4.
+- Plugin officially compatible and tested with Moodle 3.1 to 3.4.