+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [0.2.0] - 2017-11-06
+## [0.2.0] - 2017-11-09
 ### Added
 - Option to use a banner instead of a logo;
 - Option to place the navbar above or below the logo / banner.
 - Option to display of course list in Masonry grid view;
 - Option to hide site pages on front page.
+- Back to top button near lower right corner of screen.
+### Known issues
+- Favicon doesn't always work.
 
 ## [0.1.0] - 2017-11-05
 ### Added